refactoring: moving bubble_sort and dependent functions to dedicated file
