+++ src/main/c/collections/int_array/sortings/sortings.h
@@ -1,0 +1,14 @@
+#ifndef COLLECTIONS_INT_ARRAY_SORTINGS_H
+#define COLLECTIONS_INT_ARRAY_SORTINGS_H
+
+#include "collections/int_array/api.h"
+
+typedef struct IntArraySortings {
+
+    const SortIntArray BUBBLE_SORT;
+
+} IntArraySortings;
+
+IntArraySortings* int_array_sortings();
+
+#endif
