+++ src/main/c/macro/qualifiers.h
@@ -1,0 +1,6 @@
+#ifndef MACRO_QUALIFIERS_H
+#define MACRO_QUALIFIERS_H
+
+#define __pure__
+
+#endif
